requirements.txt
asgiref
dj-database-url
Django
django-cors-headers
django-environ
django-filter
djangorestframework
djangorestframework-simplejwt
Markdown
pillow
psycopg2
psycopg2-binary
PyJWT
sqlparse
typing_extensions
tzdata
